Split handler tests into individual files
