--- src/firetower/slack_app/tests/test_handlers.py
@@ -1,286 +1,0 @@
-from unittest.mock import MagicMock, call, patch
-
-import pytest
-from django.contrib.auth.models import User
-
-from firetower.auth.models import ExternalProfile, ExternalProfileType
-from firetower.incidents.models import Incident, IncidentSeverity
-from firetower.slack_app.bolt import handle_command
-from firetower.slack_app.handlers.new_incident import (
-    handle_new_command,
-    handle_new_incident_submission,
-)
-
-
-@pytest.mark.django_db
-class TestHandleInc:
-    def _make_body(self, text="", command="/ft"):
-        return {"text": text, "command": command}
-
-    def _make_command(self, command="/ft", text=""):
-        return {"command": command, "text": text}
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_help_returns_help_text(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="help")
-        command = self._make_command()
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        ack.assert_called_once()
-        respond.assert_called_once()
-        response_text = respond.call_args[0][0]
-        assert "Firetower Slack App" in response_text
-        assert "/ft help" in response_text
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_empty_text_returns_help(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="")
-        command = self._make_command()
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        ack.assert_called_once()
-        respond.assert_called_once()
-        response_text = respond.call_args[0][0]
-        assert "Firetower Slack App" in response_text
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_unknown_subcommand_returns_error(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="unknown")
-        command = self._make_command()
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        ack.assert_called_once()
-        respond.assert_called_once()
-        response_text = respond.call_args[0][0]
-        assert "Unknown command" in response_text
-        assert "/ft unknown" in response_text
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_help_uses_testinc_command(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="help", command="/ft-test")
-        command = self._make_command(command="/ft-test")
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        ack.assert_called_once()
-        response_text = respond.call_args[0][0]
-        assert "/ft-test help" in response_text
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_emits_submitted_and_completed_metrics(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="help")
-        command = self._make_command()
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        mock_statsd.increment.assert_has_calls(
-            [
-                call("slack_app.commands.submitted", tags=["subcommand:help"]),
-                call("slack_app.commands.completed", tags=["subcommand:help"]),
-            ]
-        )
-
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_emits_failed_metric_on_error(self, mock_statsd):
-        ack = MagicMock()
-        respond = MagicMock(side_effect=RuntimeError("boom"))
-        body = self._make_body(text="help")
-        command = self._make_command()
-
-        with pytest.raises(RuntimeError):
-            handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        mock_statsd.increment.assert_any_call(
-            "slack_app.commands.submitted", tags=["subcommand:help"]
-        )
-        mock_statsd.increment.assert_any_call(
-            "slack_app.commands.failed", tags=["subcommand:help"]
-        )
-
-    @patch("firetower.slack_app.bolt.get_bolt_app")
-    @patch("firetower.slack_app.bolt.statsd")
-    def test_new_subcommand_routes_correctly(self, mock_statsd, mock_get_bolt_app):
-        ack = MagicMock()
-        respond = MagicMock()
-        body = self._make_body(text="new")
-        body["trigger_id"] = "T12345"
-        command = self._make_command()
-
-        handle_command(ack=ack, body=body, command=command, respond=respond)
-
-        ack.assert_called_once()
-        mock_get_bolt_app.return_value.client.views_open.assert_called_once()
-
-
-@pytest.mark.django_db
-class TestNewIncidentModal:
-    @patch("firetower.slack_app.bolt.get_bolt_app")
-    def test_new_opens_modal(self, mock_get_bolt_app):
-        ack = MagicMock()
-        body = {"trigger_id": "T12345"}
-        command = {"text": "new"}
-        respond = MagicMock()
-
-        handle_new_command(ack, body, command, respond)
-
-        ack.assert_called_once()
-        mock_get_bolt_app.return_value.client.views_open.assert_called_once()
-        view = mock_get_bolt_app.return_value.client.views_open.call_args[1]["view"]
-        assert view["callback_id"] == "new_incident_modal"
-        assert view["type"] == "modal"
-
-
-@pytest.mark.django_db
-class TestNewIncidentSubmission:
-    def setup_method(self):
-        self.user = User.objects.create_user(
-            username="test@example.com",
-            email="test@example.com",
-            first_name="Test",
-            last_name="User",
-        )
-        ExternalProfile.objects.create(
-            user=self.user,
-            type=ExternalProfileType.SLACK,
-            external_id="U_TEST",
-        )
-
-    @patch("firetower.incidents.serializers.on_incident_created")
-    @patch("firetower.slack_app.handlers.new_incident.get_or_create_user_from_slack_id")
-    def test_creates_incident(self, mock_get_user, mock_hook):
-        mock_get_user.return_value = self.user
-
-        ack = MagicMock()
-        client = MagicMock()
-        body = {"user": {"id": "U_TEST"}}
-        view = {
-            "state": {
-                "values": {
-                    "title_block": {"title": {"value": "Test Incident"}},
-                    "severity_block": {
-                        "severity": {
-                            "selected_option": {"value": "P1"},
-                        }
-                    },
-                    "description_block": {"description": {"value": "Description"}},
-                    "private_block": {"is_private": {"selected_options": []}},
-                }
-            }
-        }
-
-        handle_new_incident_submission(ack, body, view, client)
-
-        ack.assert_called_once_with()
-        incident = Incident.objects.get(title="Test Incident")
-        assert incident.severity == IncidentSeverity.P1
-        assert incident.captain == self.user
-        assert incident.reporter == self.user
-        client.chat_postMessage.assert_called_once()
-
-    @patch("firetower.slack_app.handlers.new_incident.get_or_create_user_from_slack_id")
-    def test_validation_error(self, mock_get_user):
-        mock_get_user.return_value = self.user
-
-        ack = MagicMock()
-        client = MagicMock()
-        body = {"user": {"id": "U_TEST"}}
-        view = {
-            "state": {
-                "values": {
-                    "title_block": {"title": {"value": ""}},
-                    "severity_block": {
-                        "severity": {
-                            "selected_option": {"value": "P1"},
-                        }
-                    },
-                    "description_block": {"description": {"value": ""}},
-                    "private_block": {"is_private": {"selected_options": []}},
-                }
-            }
-        }
-
-        handle_new_incident_submission(ack, body, view, client)
-
-        ack.assert_called_once()
-        call_kwargs = ack.call_args[1]
-        assert call_kwargs["response_action"] == "errors"
-        client.chat_postMessage.assert_not_called()
-
-    @patch("firetower.slack_app.handlers.new_incident.get_or_create_user_from_slack_id")
-    def test_unknown_user_returns_error(self, mock_get_user):
-        mock_get_user.return_value = None
-
-        ack = MagicMock()
-        client = MagicMock()
-        body = {"user": {"id": "U_UNKNOWN"}}
-        view = {
-            "state": {
-                "values": {
-                    "title_block": {"title": {"value": "Test"}},
-                    "severity_block": {
-                        "severity": {"selected_option": {"value": "P1"}}
-                    },
-                    "description_block": {"description": {"value": ""}},
-                    "private_block": {"is_private": {"selected_options": []}},
-                }
-            }
-        }
-
-        handle_new_incident_submission(ack, body, view, client)
-
-        ack.assert_called_once_with()
-        client.chat_postMessage.assert_called_once()
-        msg = client.chat_postMessage.call_args[1]["text"]
-        assert "Could not identify" in msg
-
-    @pytest.fixture(autouse=False)
-    def _enable_hooks(self, settings):
-        settings.HOOKS_ENABLED = True
-
-    @pytest.mark.usefixtures("_enable_hooks")
-    @patch(
-        "firetower.incidents.serializers.on_incident_created",
-        side_effect=RuntimeError("boom"),
-    )
-    @patch("firetower.slack_app.handlers.new_incident.get_or_create_user_from_slack_id")
-    def test_save_failure_sends_error_dm(self, mock_get_user, mock_hook):
-        mock_get_user.return_value = self.user
-
-        ack = MagicMock()
-        client = MagicMock()
-        body = {"user": {"id": "U_TEST"}}
-        view = {
-            "state": {
-                "values": {
-                    "title_block": {"title": {"value": "Test Incident"}},
-                    "severity_block": {
-                        "severity": {
-                            "selected_option": {"value": "P1"},
-                        }
-                    },
-                    "description_block": {"description": {"value": "Description"}},
-                    "private_block": {"is_private": {"selected_options": []}},
-                }
-            }
-        }
-
-        handle_new_incident_submission(ack, body, view, client)
-
-        ack.assert_called_once_with()
-        client.chat_postMessage.assert_called_once()
-        msg = client.chat_postMessage.call_args[1]["text"]
-        assert "Something went wrong" in msg
-        assert "Slack channel manually" in msg
